requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
Pillow==9.5.0
pybabel==0.0.0.dev0
pydantic==1.10.8
PySocks==1.7.1
python-dotenv==1.0.0
pyzbar==0.1.9
requests==2.31.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
urllib3==2.0.2
wsproto==1.2.0
yarl==1.9.2
